Rename all card related views to a consistent format
