+++ app/src/main/java/org/jellyfin/androidtv/ui/presentation/ChannelCardPresenter.java
@@ -5,15 +5,16 @@ import android.view.ViewGroup;
 
 import androidx.leanback.widget.Presenter;
 
+import org.jellyfin.androidtv.ui.card.ChannelCardView;
 import org.jellyfin.apiclient.model.livetv.ChannelInfoDto;
 
 public class ChannelCardPresenter extends Presenter {
     class ViewHolder extends Presenter.ViewHolder {
-        private MyChannelCardView mCardView;
+        private ChannelCardView mCardView;
 
         public ViewHolder(View view) {
             super(view);
-            mCardView = (MyChannelCardView) view;
+            mCardView = (ChannelCardView) view;
         }
 
         public void setItem(ChannelInfoDto item) {
@@ -23,7 +24,7 @@ public class ChannelCardPresenter extends Presenter {
 
     @Override
     public ViewHolder onCreateViewHolder(ViewGroup parent) {
-        MyChannelCardView view = new MyChannelCardView(parent.getContext());
+        ChannelCardView view = new ChannelCardView(parent.getContext());
 
         view.setFocusable(true);
         view.setFocusableInTouchMode(true);
