+++ app/src/main/res/layout/view_card_channel.xml
