+++ app/src/main/res/layout/view_card_legacy_image.xml
