+++ app/src/main/res/layout/view_card_media_info.xml
